Added basic plugins system
