+++ frontend/src/plugins/plugins_entry.ts
@@ -1,0 +1,6 @@
+import defaultCommands from "./default_commands";
+import userCommands from "./user_commands";
+export default {
+	...defaultCommands,
+	...userCommands,
+};
