+++ frontend/src/plugins/user_commands.ts
@@ -1,0 +1,2 @@
+// You can add custom commands to extend the CLI's functionality here.
+export default {};
